+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Use reliable filesystem when updating dropin
 - Show full diagnostics with `wp redis status`
 - Always set `FS_CHMOD_FILE` when copying the `object-cache.php`
+- Call `redis_object_cache_update_dropin` during plugin update
 
 ## 2.0.3